+++ pupil_recording_interface/manager.py
@@ -9,16 +9,22 @@ from functools import reduce
 from threading import Thread
 from pathlib import Path
 
+import numpy as np
+
 from pupil_recording_interface._version import __version__
 from pupil_recording_interface.decorators import device
 from pupil_recording_interface.externals.methods import get_system_info
 from pupil_recording_interface.stream import BaseStream
-from pupil_recording_interface.utils import multiprocessing_deque, monotonic
+from pupil_recording_interface.utils import (
+    multiprocessing_deque,
+    monotonic,
+    identify_process,
+)
 
 logger = logging.getLogger(__name__)
 
 
-class StreamManager(object):
+class StreamManager:
     """ Manager for multiple streams. """
 
     def __init__(
@@ -28,7 +34,7 @@ class StreamManager(object):
         policy="new_folder",
         duration=None,
         update_interval=0.1,
-        status_timeout=5.0,
+        status_timeout=1.0,
         max_queue_size=20,
         app_info=None,
     ):
@@ -65,7 +71,7 @@ class StreamManager(object):
             Will be dropped in a future version with an asynchronous
             implementation of the update mechanism.
 
-        status_timeout: float, default 5.0
+        status_timeout: float, default 1.0
             Maximum time in seconds to wait for status updates from a stream.
             After this time, the status of the stream will be reset to the
             default.
@@ -203,6 +209,7 @@ class StreamManager(object):
                     notification_queues[stream_name],
                     priority_queues[stream_name],
                 ),
+                name=stream_name,
             )
             for stream_name, stream in streams.items()
         }
@@ -344,6 +351,7 @@ class StreamManager(object):
         self,
         key,
         format="{:.2f}",
+        nan_format="no data",
         status_dict=None,
         max_cols=None,
         sleep=None,
@@ -363,11 +371,13 @@ class StreamManager(object):
         }
 
         if len(values) > 0:
-            status_str = ", ".join(
-                f"{name}: " + format.format(value)
-                for name, value in values.items()
-                if value is not None
-            )
+            status_list = []
+            for name, value in values.items():
+                if nan_format is not None and np.isnan(value):
+                    status_list.append(f"{name}: " + nan_format.format(value))
+                elif value is not None:
+                    status_list.append(f"{name}: " + format.format(value))
+            status_str = ", ".join(status_list)
         else:
             return None
 
@@ -431,6 +441,8 @@ class StreamManager(object):
             logger.debug(f"Streaming for {self.duration} seconds")
         logger.debug(f"Run start time: {self._start_time}")
         logger.debug(f"Run start time monotonic: {self._start_time_monotonic}")
+
+        identify_process("manager")
 
     def _update(self):
         """ Update status and notify streams. """